+++ packages/terra-framework-docs/src/terra-dev-site/test/layout/LayoutDefault.test.jsx
@@ -1,16 +1,18 @@
 import React from 'react';
 import classNames from 'classnames/bind';
 import Layout from 'terra-layout';
-import TestMenu from '../common/TestMenu';
-import TestContent from '../common/TestContent';
+import TestHeader from './common/TestHeader';
+import TestMenu from './common/TestMenu';
+import TestContent from './common/TestContent';
 import styles from './TestLayoutCommon.module.scss';
 
 const cx = classNames.bind(styles);
 
-const LayoutNoHeader = () => (
+const LayoutDefault = () => (
   <Layout
     className={cx('layout-test')}
-    key="layout-no-header"
+    key="layout-default"
+    header={<TestHeader />}
     menu={<TestMenu />}
     menuText="Menu Text"
     id="test-root"
@@ -19,4 +21,4 @@ const LayoutNoHeader = () => (
   </Layout>
 );
 
-export default LayoutNoHeader;
+export default LayoutDefault;
