+++ packages/terra-framework-docs/src/terra-dev-site/test/layout/LayoutNoHeader.test.jsx
@@ -1,21 +1,22 @@
 import React from 'react';
 import classNames from 'classnames/bind';
 import Layout from 'terra-layout';
-import TestHeader from '../common/TestHeader';
-import TestContent from '../common/TestContent';
+import TestMenu from './common/TestMenu';
+import TestContent from './common/TestContent';
 import styles from './TestLayoutCommon.module.scss';
 
 const cx = classNames.bind(styles);
 
-const LayoutNoMenu = () => (
+const LayoutNoHeader = () => (
   <Layout
     className={cx('layout-test')}
-    key="layout-no-menu"
-    header={<TestHeader />}
+    key="layout-no-header"
+    menu={<TestMenu />}
+    menuText="Menu Text"
     id="test-root"
   >
     <TestContent />
   </Layout>
 );
 
-export default LayoutNoMenu;
+export default LayoutNoHeader;
