+++ packages/terra-framework-docs/src/terra-dev-site/test/layout/LayoutNoMenu.test.jsx
@@ -1,24 +1,21 @@
 import React from 'react';
 import classNames from 'classnames/bind';
 import Layout from 'terra-layout';
-import TestHeader from '../common/TestHeader';
-import TestMenu from '../common/TestMenu';
-import TestContent from '../common/TestContent';
+import TestHeader from './common/TestHeader';
+import TestContent from './common/TestContent';
 import styles from './TestLayoutCommon.module.scss';
 
 const cx = classNames.bind(styles);
 
-const LayoutDefault = () => (
+const LayoutNoMenu = () => (
   <Layout
     className={cx('layout-test')}
-    key="layout-default"
+    key="layout-no-menu"
     header={<TestHeader />}
-    menu={<TestMenu />}
-    menuText="Menu Text"
     id="test-root"
   >
     <TestContent />
   </Layout>
 );
 
-export default LayoutDefault;
+export default LayoutNoMenu;
