+++ packages/terra-framework-docs/src/terra-dev-site/test/layout/LayoutSlidePanelExample.test.jsx
